+++ src/com/intellij/idea/plugin/hybris/psi/injector/GroovyLanguageInjector.kt
@@ -17,6 +17,7 @@
  */
 package com.intellij.idea.plugin.hybris.psi.injector
 
+import com.intellij.idea.plugin.hybris.impex.psi.ImpexGroovyScriptBody
 import com.intellij.idea.plugin.hybris.impex.psi.ImpexString
 import com.intellij.idea.plugin.hybris.system.type.ScriptType
 import com.intellij.openapi.diagnostic.Logger
@@ -47,24 +48,31 @@ class GroovyLanguageInjector : LanguageInjector {
         val xmlFile = host.containingFile as? XmlFile
             ?: return
 
-        LanguageInjectionUtil.tryInject(xmlFile, host, ScriptType.GROOVY) { length, offset -> injectLanguage(injectionPlacesRegistrar, length, offset) }
+        LanguageInjectionUtil.tryInject(xmlFile, host, ScriptType.GROOVY) { length, offset ->
+            injectLanguage(injectionPlacesRegistrar, length, offset)
+        }
     }
 
     private fun handleImpex(host: PsiLanguageInjectionHost, injectionPlacesRegistrar: InjectedLanguagePlaces) {
-        val impexString = host as? ImpexString
-            ?: return
+        when (host) {
+            is ImpexString -> {
+                val hostString = StringUtil.unquoteString(host.getText()).lowercase()
+                if (StringUtil.trim(hostString).replaceFirst("\"", "").startsWith(groovyMarker)) {
+                    val markerOffset = setOf("beforeeach:", "aftereach:", "if:")
+                        .map { it to host.getText().indexOf(it, 0, true) }
+                        .firstOrNull { it.second > -1 }
+                        ?.let { it.first.length + it.second }
+                        ?: offset
 
-        val hostString = StringUtil.unquoteString(impexString.text).lowercase()
-        if (StringUtil.trim(hostString).replaceFirst("\"", "").startsWith(groovyMarker)) {
-            val markerOffset = setOf("beforeeach:", "aftereach:", "if:")
-                .map { it to impexString.text.indexOf(it, 0, true) }
-                .firstOrNull { it.second > -1 }
-                ?.let { it.first.length + it.second }
-                ?: offset
+                    injectLanguage(injectionPlacesRegistrar, host.getTextLength() - markerOffset - quoteSymbolLength, markerOffset)
+                } else if (LanguageInjectionUtil.getLanguageForInjection(host) == ScriptType.GROOVY) {
+                    injectLanguage(injectionPlacesRegistrar, host.getTextLength() - quoteSymbolLength - 1, quoteSymbolLength)
+                }
+            }
 
-            injectLanguage(injectionPlacesRegistrar, impexString.textLength - markerOffset - quoteSymbolLength, markerOffset)
-        } else if (LanguageInjectionUtil.getLanguageForInjection(impexString) == ScriptType.GROOVY) {
-            injectLanguage(injectionPlacesRegistrar, impexString.textLength - quoteSymbolLength - 1, quoteSymbolLength)
+            is ImpexGroovyScriptBody -> {
+                injectLanguage(injectionPlacesRegistrar, host.getTextLength(), 0)
+            }
         }
     }